First draft saved for the project
The model presented makes the assumption that there exists a joint latent feature space F such that user-item interactions are inner products in that space.  The strength
of this model is to predict a probable grade that user u would give to films if she ever watched it. Thus, we can recommend her films that she has not watched but that she may enjoy.

The model is modeled as a multi-variable optimization problem, which is solved by taylor-based line search.